test(custom_op): 将测试张量数据类型更改为bfloat16
为了测试fused_attention函数在bfloat16数据类型下的表现，将测试张量q、k、v的数据类型从float32更改为bfloat16

+++ custom_op/test_attentions.py
@@ -24,9 +24,9 @@ def fused_attention(q: torch.Tensor, k: torch.Tensor, v: torch.Tensor) -> torch.
 def main():
     device = "cuda" if torch.cuda.is_available() else "cpu"
     print(f"device: {device}")
-    q = torch.randn((32, 32), dtype=torch.float32, device=device)
-    k = torch.randn((32, 32), dtype=torch.float32, device=device)
-    v = torch.randn((32, 32), dtype=torch.float32, device=device)
+    q = torch.randn((32, 32), dtype=torch.bfloat16, device=device)
+    k = torch.randn((32, 32), dtype=torch.bfloat16, device=device)
+    v = torch.randn((32, 32), dtype=torch.bfloat16, device=device)
     print(fused_attention(q, k, v))
